Add files via upload
Hydrogen Release and Uptake by Soils (HORUS) model uses the existing temperature and moisture functions and explicitly incorporates a description of the impacts of the composition of soil organic matter on H2 uptake. The HORUS model apply of the RothC model to estimate microbial activity, then applies a conversion factor to translate it into H2 uptake potential.